auto use fake token for local env

+++ src/main/java/com/dtsx/astra/cli/operations/config/ConfigCreateOperation.java
@@ -16,18 +16,19 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.function.Function;
 
 @RequiredArgsConstructor
 public class ConfigCreateOperation implements Operation<ConfigCreateResult> {
     private final CliContext ctx;
     private final AstraConfig config;
-    private final OrgGateway orgGateway;
+    private final Function<AstraToken, OrgGateway> orgGateway;
     private final OrgGateway.Stateless statelessOrgGateway;
     private final CreateConfigRequest request;
 
     public record CreateConfigRequest(
         Optional<ProfileName> profileName,
-        AstraToken token,
+        Optional<AstraToken> token,
         AstraEnvironment env,
         Optional<Boolean> overwrite,
         boolean setDefault,
@@ -40,13 +41,13 @@ public class ConfigCreateOperation implements Operation<ConfigCreateResult> {
     public record ProfileCreated(ProfileName profileName, boolean overwritten, boolean isDefault) implements ConfigCreateResult {}
     public record ProfileIllegallyExists(ProfileName profileName) implements ConfigCreateResult {}
     public record ViolatedFailIfExists() implements ConfigCreateResult {}
+    public record MissingToken() implements ConfigCreateResult {}
     public record InvalidToken(Optional<AstraEnvironment> hint) implements ConfigCreateResult {}
     public record NameRequiredIfNotValidated() implements ConfigCreateResult {}
 
     @Override
     public ConfigCreateResult execute() {
-        val maybeProfileName = validateTokenAndResolveName(orgGateway, request);
-
+        val maybeProfileName = validateTokenAndResolveName(request);
         if (maybeProfileName.isLeft()) {
             return maybeProfileName.getLeft();
         }
@@ -62,13 +63,17 @@ public class ConfigCreateOperation implements Operation<ConfigCreateResult> {
             }
         }
 
+        val maybeToken = resolveToken(request.token, request.env);
+        if (maybeToken.isLeft()) {
+            return maybeToken.getLeft();
+        }
+
         config.modify((ctx) -> {
             ctx.deleteProfile(profileName);
-            ctx.createProfile(profileName, request.token, request.env, request.localEndpoint);
+            ctx.createProfile(profileName, maybeToken.getRight(), request.env, request.localEndpoint);
 
             if (request.setDefault) {
-                ctx.deleteProfile(ProfileName.DEFAULT);
-                ctx.createProfile(ProfileName.DEFAULT, request.token, request.env, request.localEndpoint);
+                ctx.copyProfile(config.lookupProfile(profileName).orElseThrow(), ProfileName.DEFAULT);
             }
         });
 
@@ -79,17 +84,19 @@ public class ConfigCreateOperation implements Operation<ConfigCreateResult> {
         );
     }
 
-    private Either<ConfigCreateResult, ProfileName> validateTokenAndResolveName(OrgGateway orgGateway, CreateConfigRequest request) {
-        if (!request.validate) {
+    private Either<ConfigCreateResult, ProfileName> validateTokenAndResolveName(CreateConfigRequest request) {
+        if (!request.validate || request.token.isEmpty()) {
             return Either.fromOptional(request.profileName, NameRequiredIfNotValidated::new);
         }
 
+        val token = request.token.get();
+
         return ctx.log().loading("Validating your Astra token", (_) -> {
             try {
-                val name = ProfileName.mkUnsafe(orgGateway.current().getName());
+                val name = ProfileName.mkUnsafe(orgGateway.apply(token).current().getName());
                 return Either.pure(name);
             } catch (AuthenticationException e) {
-                val validEnv = statelessOrgGateway.resolveOrganizationEnvironment(request.token).map(Pair::getLeft);
+                val validEnv = statelessOrgGateway.resolveOrganizationEnvironment(token).map(Pair::getLeft);
                 return Either.left(new InvalidToken(validEnv));
             } catch (Exception e) {
                 ctx.log().hint("You can use @!--no-validate!@ to skip token validation and create a profile with an explicit name.");
@@ -109,5 +116,19 @@ public class ConfigCreateOperation implements Operation<ConfigCreateResult> {
         return (!shouldOverwrite)
             ? Optional.of(new ProfileIllegallyExists(profileName))
             : Optional.empty();
+    }
+
+    private Either<ConfigCreateResult, AstraToken> resolveToken(Optional<AstraToken> token, AstraEnvironment env) {
+        if (token.isPresent()) {
+            return Either.pure(token.get());
+        }
+
+        if (env.isLocal()) {
+            ctx.log().info("@|italic Using a fake astra token for your local environment.|@");
+            ctx.log().info("@|italic Provide an explicit token if using an actual auth manager.|@");
+            return Either.pure(AstraToken.FAKE);
+        }
+
+        return Either.left(new MissingToken());
     }
 }
